Fixed loop indentations and moved FollowsNodeException
