+++ generator/src/main/java/org/RaceGuard/Exceptions/FollowsNodeException.java
@@ -1,4 +1,4 @@
-package org.RaceGuard.AstNodes;
+package org.RaceGuard.Exceptions;
 
 public class FollowsNodeException extends RuntimeException{
     public FollowsNodeException(String message){
